requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiosqlite==0.19.0
    # via prefect
alembic==1.10.4
    # via prefect
anyio==3.6.2
    # via
    #   httpcore
    #   prefect
    #   starlette
apprise==1.3.0
    # via prefect
asgi-lifespan==2.1.0
    # via prefect
asyncpg==0.27.0
    # via prefect
attrs==23.1.0
    # via jsonschema
boto3==1.24.96
    # via -r requirements.in
botocore==1.27.96
    # via
    #   boto3
    #   s3transfer
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via
    #   apprise
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   apprise
    #   prefect
    #   typer
    #   uvicorn
cloudpickle==2.2.1
    # via prefect
colorama==0.4.6
    # via griffe
coolname==2.2.0
    # via prefect
croniter==1.3.14
    # via prefect
cryptography==40.0.2
    # via prefect
dateparser==1.1.8
    # via prefect
docker==6.0.1
    # via prefect
exceptiongroup==1.1.1
    # via pytest
fastapi==0.95.1
    # via prefect
fsspec==2023.4.0
    # via prefect
gdal==3.6.4
    # via -r requirements.in
google-auth==2.17.3
    # via kubernetes
greenlet==2.0.2
    # via sqlalchemy
griffe==0.27.1
    # via prefect
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.17.0
    # via httpx
httpx[http2]==0.24.0
    # via prefect
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via prefect
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.32
    # via prefect
jsonpointer==2.3
    # via jsonpatch
jsonschema==4.17.3
    # via prefect
kubernetes==26.1.0
    # via prefect
mako==1.2.4
    # via alembic
markdown==3.4.3
    # via apprise
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
numpy==1.24.3
    # via
    #   -r requirements.in
    #   pandas
oauthlib==3.2.2
    # via requests-oauthlib
orjson==3.8.10
    # via prefect
packaging==23.1
    # via
    #   docker
    #   prefect
    #   pytest
pandas==1.5.3
    # via -r requirements.in
pathspec==0.11.1
    # via prefect
pendulum==2.1.2
    # via prefect
pluggy==1.0.0
    # via pytest
prefect==2.10.6
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via
    #   fastapi
    #   prefect
pygments==2.15.1
    # via rich
pyrsistent==0.19.3
    # via jsonschema
pytest==7.3.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   croniter
    #   dateparser
    #   kubernetes
    #   pandas
    #   pendulum
python-dotenv==0.21.1
    # via -r requirements.in
python-slugify==8.0.1
    # via prefect
pytz==2023.3
    # via
    #   dateparser
    #   pandas
    #   prefect
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via
    #   apprise
    #   kubernetes
    #   prefect
readchar==4.0.5
    # via prefect
regex==2023.3.23
    # via dateparser
requests==2.29.0
    # via
    #   apprise
    #   docker
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   apprise
    #   kubernetes
rich==13.3.5
    # via prefect
rsa==4.9
    # via google-auth
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   asgi-lifespan
    #   httpcore
    #   httpx
sqlalchemy[asyncio]==1.4.48
    # via
    #   -r requirements.in
    #   alembic
    #   prefect
starlette==0.26.1
    # via fastapi
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via prefect
tomli==2.0.1
    # via pytest
typer==0.7.0
    # via prefect
typing-extensions==4.5.0
    # via
    #   alembic
    #   prefect
    #   pydantic
tzdata==2023.3
    # via pytz-deprecation-shim
tzlocal==4.3
    # via dateparser
urllib3==1.26.15
    # via
    #   botocore
    #   docker
    #   kubernetes
    #   requests
uvicorn==0.22.0
    # via prefect
websocket-client==1.5.1
    # via
    #   docker
    #   kubernetes
websockets==11.0.2
    # via prefect

# The following packages are considered to be unsafe in a requirements file:
# setuptools
